+++ src/Arborescence.Models.Specialized/Incidence/Int32IncidenceGraph.Factory.cs
@@ -7,11 +7,7 @@ namespace Arborescence.Models.Specialized
     using System.Runtime.InteropServices;
 #endif
 
-    /// <summary>
-    /// Provides a set of initialization methods for instances
-    /// of the <see cref="Int32IncidenceGraph"/> type.
-    /// </summary>
-    public static class Int32IncidenceGraphFactory
+    partial struct Int32IncidenceGraph
     {
         /// <summary>
         /// Creates an <see cref="Int32IncidenceGraph"/> with the specified edges defined by their endpoint pairs.
@@ -66,13 +62,13 @@ namespace Arborescence.Models.Specialized
                 ArgumentOutOfRangeExceptionHelpers.Throw(nameof(vertexCount));
 
             // We cannot reuse this overload taking span:
-            // Int32IncidenceGraphFactory.FromEdgesUnchecked(int, Span<Endpoints<int>>)
+            // Int32IncidenceGraph.FromEdgesUnchecked(int, Span<Endpoints<int>>)
             // Because the sorting algorithm that takes spans is only available in .NET 5 and later:
             // MemoryExtensions.Sort<TKey,TValue,TComparer>(Span<TKey>, Span<TValue>, TComparer)
             // https://learn.microsoft.com/en-us/dotnet/api/system.memoryextensions.sort?view=net-6.0
 
             // We need to implement sorting independently with another overload that takes arrays:
-            // Int32IncidenceGraphFactory.FromEdgesUnchecked(int, Endpoints<int>[])
+            // Int32IncidenceGraph.FromEdgesUnchecked(int, Endpoints<int>[])
             // Because the sorting algorithm that takes arrays is available on older versions of .NET:
             // Array.Sort<TKey,TValue>(TKey[], TValue[], IComparer<TKey>)
             // https://learn.microsoft.com/en-us/dotnet/api/system.array.sort?view=net-6.0